src/pyproject.toml
[tool.poetry]
name = "Bigpay"
version = "1.0.0"
description = "download price index"
authors = ["Omar"]

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.26.0"
psycopg2-binary = "^2.9.2"


[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
